Move system admin commands to their own folder in commands directory.
